requirements.txt
boto3
botocore
s3transfer
jmespath
python-dateutil
urllib3
six
awscli
python-dotenv